+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 * `destroyResource` no longer ignores exceptions thrown from resource releasing
   functions.
 * Change the default number of stripes to 1.
+* Do not exceed the maximum number of resources if the number of stripes does
+  not divide it.
 
 # resource-pool-0.4.0.0 (2023-01-16)
 * Require `poolMaxResources` to be not smaller than the number of stripes.